Set maximum tokenized line length depending of last statement.
The Turbo Basic XL interpreter has a bug regarding lines of exactly 255
bytes of tokenized length, with certain statements it fails to correctly
detect the end of the last statement.

So, adds a function that returns the maximum length depending on the
last statement and use that value instead of a fixed one.

Signed-off-by: Daniel Serpell <daniel.serpell@gmail.com>